rag/requirements.txt
flask
pydantic
python-dotenv
openai
